# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GigWork — Gig-based local work marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List active jobs (filterable by city, category, work type)
      security:
        - bearerAuth: []
      parameters:
        - name: city
          in: query
          schema:
            type: string
        - name: category_id
          in: query
          schema:
            type: ["integer", "null"]
        - name: work_type
          in: query
          schema:
            type: ["string", "null"]
            enum: [hourly, daily, monthly, null]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Jobs list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"

    post:
      operationId: createJob
      tags: [jobs]
      summary: Post a new job (provider only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          description: Validation error or scam detected
        "403":
          description: Provider role required
  /v1/jobs/search:
    get:
      operationId: searchJobs
      tags: [jobs]
      summary: Full text search jobs
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"
  /v1/jobs/my:
    get:
      operationId: listMyJobs
      tags: [jobs]
      summary: List jobs posted by current provider
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: [draft, active, paused, filled, expired, removed, null]
      responses:
        "200":
          description: My posted jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
  /v1/jobs/{jobId}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get full job detail
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Job not found

    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job (provider only, own jobs)
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "403":
          description: Not your job
        "404":
          description: Job not found

    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Remove a job (soft delete)
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Job removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /v1/jobs/{jobId}/report:
    post:
      operationId: reportJob
      tags: [jobs]
      summary: Report a suspicious job
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "200":
          description: Report submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ── APPLICATIONS ──────────────────────────────────────────────────────────────
...
  /v1/reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review after job completion
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "400":
          description: Already reviewed or not eligible
  /v1/users/{userId}/reviews:
    get:
      operationId: listUserReviews
      tags: [reviews]
      summary: List reviews for a user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reviews list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

  # ── CHAT ──────────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── AUTH SCHEMAS ─────────────────────────────────────────────────────────────
...
    UserProfile:
      type: object
      required: [id, phone, current_role, created_at]
      properties:
        id:
          type: string
        phone:
          type: string
        name:
          type: ["string", "null"]
        profile_photo_url:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        current_role:
          type: string
          enum: [seeker, provider]
        is_phone_verified:
          type: boolean
        is_id_verified:
          type: boolean
        trust_score:
          type: number
        is_banned:
          type: boolean
        seeker_profile:
          $ref: "#/components/schemas/SeekerProfile"
        provider_profile:
          $ref: "#/components/schemas/ProviderProfile"
        created_at:
          type: string
    SeekerProfile:
      type: object
      properties:
        bio:
          type: ["string", "null"]
        availability:
          type: string
          enum: [immediate, within_week, flexible]
        preferred_work_type:
          type: string
          enum: [hourly, daily, monthly, any]
        salary_expectation:
          type: ["integer", "null"]
        years_experience:
          type: integer
        total_jobs_completed:
          type: integer
        avg_rating:
          type: number
        total_ratings:
          type: integer
        attendance_score:
          type: number
        skills:
          type: array
          items:
            type: string
    ProviderProfile:
      type: object
      properties:
        business_name:
          type: ["string", "null"]
        business_type:
          type: ["string", "null"]
        total_jobs_posted:
          type: integer
        total_hires:
          type: integer
        avg_rating:
          type: number
        total_ratings:
          type: integer
        avg_response_hours:
          type: ["number", "null"]
        hiring_success_rate:
          type: number
...
    Category:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
        name_hi:
          type: ["string", "null"]
        icon_name:
          type: ["string", "null"]
        sort_order:
          type: integer

    # ── JOB SCHEMAS ──────────────────────────────────────────────────────────────
    Job:
      type: object
      required: [id, title, work_type, city, status, created_at]
      properties:
        id:
          type: string
        provider_id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        category_id:
          type: ["integer", "null"]
        category:
          $ref: "#/components/schemas/Category"
        work_type:
          type: string
          enum: [hourly, daily, monthly]
        # hourly
        hourly_rate:
          type: ["integer", "null"]
        total_hours:
          type: ["integer", "null"]
        shift_start_time:
          type: ["string", "null"]
        shift_end_time:
          type: ["string", "null"]
        same_day_payment:
          type: ["boolean", "null"]
        urgent_hiring:
          type: ["boolean", "null"]
        # daily
        daily_wage:
          type: ["integer", "null"]
        num_working_days:
          type: ["integer", "null"]
        food_included:
          type: ["boolean", "null"]
        accommodation_included:
          type: ["boolean", "null"]
        overtime_available:
          type: ["boolean", "null"]
        # monthly
        monthly_salary:
          type: ["integer", "null"]
        joining_date:
          type: ["string", "null"]
        experience_required:
          type: ["integer", "null"]
        working_days_per_week:
          type: ["integer", "null"]
        pf_esi_included:
          type: ["boolean", "null"]
        salary_negotiable:
          type: ["boolean", "null"]
        # location
        city:
          type: string
        address:
          type: ["string", "null"]
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        # status
        status:
          type: string
          enum: [draft, active, paused, filled, expired, removed]
        is_urgent:
          type: boolean
        is_verified:
          type: boolean
        views_count:
          type: integer
        applications_count:
          type: integer
        skills_required:
          type: array
          items:
            type: string
        provider:
          $ref: "#/components/schemas/UserProfile"
        expires_at:
          type: ["string", "null"]
        created_at:
          type: string
        has_applied:
          type: ["boolean", "null"]
    JobInput:
      type: object
      required: [title, work_type, city]
      properties:
        title:
          type: string
        category_id:
          type: ["integer", "null"]
        description:
          type: string
        work_type:
          type: string
          enum: [hourly, daily, monthly]
        hourly_rate:
          type: ["integer", "null"]
        total_hours:
          type: ["integer", "null"]
        shift_start_time:
          type: ["string", "null"]
        shift_end_time:
          type: ["string", "null"]
        same_day_payment:
          type: ["boolean", "null"]
        urgent_hiring:
          type: ["boolean", "null"]
        daily_wage:
          type: ["integer", "null"]
        num_working_days:
          type: ["integer", "null"]
        food_included:
          type: ["boolean", "null"]
        accommodation_included:
          type: ["boolean", "null"]
        overtime_available:
          type: ["boolean", "null"]
        monthly_salary:
          type: ["integer", "null"]
        joining_date:
          type: ["string", "null"]
        experience_required:
          type: ["integer", "null"]
        working_days_per_week:
          type: ["integer", "null"]
        pf_esi_included:
          type: ["boolean", "null"]
        salary_negotiable:
          type: ["boolean", "null"]
        city:
          type: string
        address:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        is_urgent:
          type: boolean
        skills_required:
          type: array
          items:
            type: string
    JobUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        daily_wage:
          type: ["integer", "null"]
        hourly_rate:
          type: ["integer", "null"]
        monthly_salary:
          type: ["integer", "null"]
        is_urgent:
          type: boolean
        status:
          type: string
          enum: [draft, active, paused, filled, expired, removed]
        food_included:
          type: ["boolean", "null"]
        accommodation_included:
          type: ["boolean", "null"]
    JobListResponse:
      type: object
      required: [jobs, total, page, limit]
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        has_more:
          type: boolean
    ReportInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          enum: [fake_salary, scam, spam, inappropriate, already_filled, other]
        description:
          type: string

    # ── APPLICATION SCHEMAS ───────────────────────────────────────────────────────
...
    Review:
      type: object
      required: [id, reviewer_id, reviewee_id, rating, reviewer_role, created_at]
      properties:
        id:
          type: string
        application_id:
          type: ["string", "null"]
        reviewer_id:
          type: string
        reviewee_id:
          type: string
        reviewer_role:
          type: string
          enum: [seeker, provider]
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        reviewer:
          $ref: "#/components/schemas/UserProfile"
        created_at:
          type: string
    ReviewInput:
      type: object
      required: [application_id, reviewee_id, rating, reviewer_role]
      properties:
        application_id:
          type: string
        reviewee_id:
          type: string
        reviewer_role:
          type: string
          enum: [seeker, provider]
        rating:
          type: integer
        comment:
          type: string

    # ── CHAT SCHEMAS ──────────────────────────────────────────────────────────────
...